+++ tests/integration/copy-assets/tests/webpack.test.ts
@@ -1,7 +1,0 @@
-import { testPublicHtml } from './testPublicHtml';
-
-describe('copy assets', () => {
-  test(`should copy public html and replace the assetPrefix variable in webpack`, async () => {
-    await testPublicHtml('webpack');
-  });
-});