+++ tests/integration/esbuild/fixtures/legacy-minify-js/modern.config.ts
@@ -1,8 +1,0 @@
-import { appTools, defineLegacyConfig } from '@modern-js/app-tools';
-
-export default defineLegacyConfig({
-  tools: {
-    esbuild: {},
-  },
-  plugins: [appTools()],
-});