+++ tests/integration/load-config/fixtures/async-config-function/modern.config.ts
@@ -1,12 +1,11 @@
-import { appTools, defineConfig } from '@modern-js/app-tools';
+import { applyBaseConfig } from '../../../../utils/applyBaseConfig';
 
-export default defineConfig(async () => {
-  return {
+export default async () => {
+  return applyBaseConfig({
     output: {
       distPath: {
         root: 'dist/foo',
       },
     },
-    plugins: [appTools()],
-  };
-});
+  });
+};