+++ tests/integration/load-config/fixtures/local-config-function/modern.config.ts
@@ -1,10 +1,9 @@
-import { appTools, defineConfig } from '@modern-js/app-tools';
+import { applyBaseConfig } from '../../../../utils/applyBaseConfig';
 
-export default defineConfig({
+export default applyBaseConfig({
   output: {
     distPath: {
       root: 'dist/foo',
     },
   },
-  plugins: [appTools()],
 });