+++ packages/builder/plugin-esbuild/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @modern-js/builder-plugin-esbuild
 
+## 2.35.0
+
+### Patch Changes
+
+- Updated dependencies [225f065]
+- Updated dependencies [0300daf]
+- Updated dependencies [f509548]
+- Updated dependencies [28128ca]
+- Updated dependencies [4027a6a]
+- Updated dependencies [3d93e81]
+  - @modern-js/builder-shared@2.35.0
+
 ## 2.34.0
 
 ### Patch Changes