+++ packages/devtools/client/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/devtools-client
 
+## 2.35.0
+
+### Patch Changes
+
+- Updated dependencies [d87a426]
+  - @modern-js/runtime@2.35.0
+  - @modern-js/devtools-kit@2.35.0
+  - @modern-js/types@2.35.0
+
 ## 2.34.0
 
 ### Patch Changes