+++ packages/generator/generator-plugin/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @modern-js/generator-plugin
 
+## 3.2.1
+
+### Patch Changes
+
+- Updated dependencies [15b834f]
+  - @modern-js/utils@2.35.0
+  - @modern-js/plugin-i18n@2.35.0
+  - @modern-js/generator-common@3.2.1
+  - @modern-js/generator-utils@3.2.1
+  - @modern-js/new-action@2.35.0
+
 ## 3.2.0
 
 ### Patch Changes