+++ packages/toolkit/upgrade/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @modern-js/upgrade
 
+## 2.35.0
+
+### Patch Changes
+
+- Updated dependencies [15b834f]
+  - @modern-js/utils@2.35.0
+  - @modern-js/plugin-i18n@2.35.0
+
 ## 2.34.0
 
 ### Patch Changes